// packages/data-service-generator/src/server/resource/dto/entity-dto-type-enum.ts
export enum EntityDtoTypeEnum {
  Entity = "entity",
  CreateInput = "createInput",
  UpdateInput = "updateInput",
  WhereInput = "whereInput",
  WhereUniqueInput = "whereUniqueInput",

  OrderByInput = "orderByInput",

  CreateArgs = "createArgs",
  DeleteArgs = "deleteArgs",
  FindMany = "findMany",
  FindUnique = "findUnique",
  UpdateArgs = "updateArgs",

  RelationCreateNestedManyWithoutSourceInput = "relationCreateNestedManyWithoutSourceInput",
  RelationUpdateManyWithoutSourceInput = "relationUpdateManyWithoutSourceInput",
  ListRelationFilter = "listRelationFilter",
}